Converted digit view to a Shape
Our underlying view hierarchy was bloated because we had a view _per-bar_ per view. So 7 * digits + (3 * 3 for endcaps and middle). This was something like 93 views for a UPC-A barcode and 100 for EAN13. There is room for further improvement here but this change converts `BarcodeDigitView` to a `Shape` and instead describes the filled portion via paths. This an order of magnitude reduction in views: just one for each digit and 3 for endcaps + middle. This means 15 for UPC-A and 16 for EAN13. Nice!

Updated README to remove reference to this conversion since it's done :)

Removed personal xcode preferences from the sample app.